moved get ingredients and get recipes by ingredient to the corresponding controller files
